Separate local context that conforms to local type registry
