+++ generators/swift/base/src/index.ts
@@ -1,4 +1,4 @@
+export * from "./AsIs";
 export * from "./cli";
 export * from "./context";
 export * from "./project";
-export * from "./AsIs";
